IMAP fetch and store commands.

+++ crates/imap-proto/src/protocol/fetch.rs
@@ -140,7 +140,7 @@ pub enum DataItem<'x> {
         contents: Option<Cow<'x, str>>,
     },
     ModSeq {
-        modseq: u32,
+        modseq: u64,
     },
     EmailId {
         email_id: String,
